Migrations adjustment and pm2

+++ src/database/knex/migrations/20231001220142_ingredientsTags.js
@@ -1,9 +1,17 @@
-exports.up = knex => knex.schema.createTable( "ingredients", table =>{
-  table.increments("id");
-  table.text("tags").nullable();
-  table.integer("dish_id").references("id").inTable("dishs").onDelete("CASCADE")
 
-}) 
+  exports.up = async function (knex) {
+    const tableExists = await knex.schema.hasTable('ingredients');
+
+    if (!tableExists) {
+      return knex.schema.createTable('ingredients', function (table) {
+      table.increments("id");
+      table.text("tags").nullable();
+      table.integer("dish_id").references("id").inTable("dishs").onDelete("CASCADE")
+      })
+    }
+}
 
 
-exports.down = knex => knex.schema.dropTable("ingredients")
+exports.down = function (knex) {
+  return knex.schema.dropTableIfExists('ingredients');
+};
